Refactor workspace manager
Refactor workspace manager into smaller components for better organization and maintainability.
