+++ pyside6_overlay/ui/components/__init__.py
@@ -8,10 +8,16 @@ from .header_ui import HeaderUI
 from .control_panel_ui import ControlPanelUI
 from .rf4s_control_ui import RF4SControlUI
 from .status_bar_ui import StatusBarUI
+from .workspace_layout_selector import WorkspaceLayoutSelector
+from .workspace_navigation import WorkspaceNavigation
+from .workspace_content import WorkspaceContent
 
 __all__ = [
     'HeaderUI',
     'ControlPanelUI', 
     'RF4SControlUI',
-    'StatusBarUI'
+    'StatusBarUI',
+    'WorkspaceLayoutSelector',
+    'WorkspaceNavigation',
+    'WorkspaceContent'
 ]
